requirements.txt
Faker
pandas
pyarrow
boto3
schedule
fastapi
uvicorn
python-dotenv
sqlalchemy
pyodbc
duckdb
